SD Refund 로직 수정 및 자료형 수정

+++ src/main/java/com/youthcase/orderflow/pr/dto/PRRecommendItemDto.java
@@ -2,11 +2,13 @@ package com.youthcase.orderflow.pr.dto;
 
 import com.youthcase.orderflow.master.product.domain.StorageMethod;
 
+import java.math.BigDecimal;
+
 public record PRRecommendItemDto(
         String productCode,
         String productName,
-        Integer suggestedQty,
-        Integer unitPrice,
+        Long suggestedQty,
+        BigDecimal unitPrice,
         StorageMethod storageMethod,
         String category,
         String reason,
